backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy>=1.24.0,<2.0.0
Pillow>=10.0.0
pytesseract==0.3.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
scikit-learn==1.3.2
pandas==2.1.3
cryptography>=41.0.0
aiofiles==23.2.1
cmake>=3.25.0
dlib>=19.24.0
face-recognition>=1.3.0

# AI Agent System (NEW)
crewai>=0.1.0
crewai-tools>=0.1.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.1
openai>=1.0.0

